Move Message to separate module
