+++ src/lib.rs
@@ -1,3 +1,4 @@
 mod error;
+mod message;
 mod route;
 pub mod service;
